La till modell på product och role. Inclusive uppdaterat dessas controllers

+++ src/main/java/se/iths/grocerylist/controller/UserController.java
@@ -6,12 +6,13 @@ import org.springframework.web.bind.annotation.*;
 import se.iths.grocerylist.entity.UserEntity;
 import se.iths.grocerylist.exception.BadRequestException;
 import se.iths.grocerylist.exception.EntityNotFoundException;
+import se.iths.grocerylist.exception.MethodNotAllowedException;
 import se.iths.grocerylist.mapper.UserMapper;
 import se.iths.grocerylist.model.UserModel;
 import se.iths.grocerylist.sender.Sender;
 import se.iths.grocerylist.service.UserService;
 
-import java.util.ArrayList;
+
 import java.util.Optional;
 
 @RestController
@@ -29,20 +30,21 @@ public class UserController {
     }
 
     @PostMapping("signup")
-    public ResponseEntity<UserModel> createUser(@ModelAttribute("signup") @RequestBody UserModel user){
+    public ResponseEntity<UserModel> createUser( /*@ModelAttribute("signup")*/ @RequestBody UserModel user ){
+
 
         if(user.getUsername()==null || user.getUsername().isEmpty()){
-            throw new BadRequestException("Empty Username");
+            throw new BadRequestException("Username cannot be empty" );
         }
 
         if(user.getEmail()==null || user.getEmail().isEmpty()) {
 
-            throw new BadRequestException("Empty Email");
+            throw new BadRequestException("Email cannot be empty");
         }
 
         if(user.getPassword()==null || user.getPassword().isEmpty()) {
 
-            throw new BadRequestException("Empty Password");
+            throw new BadRequestException("Password cannot be empty");
         }
 
 //        sender.sendMessage(user.getUsername());
@@ -51,6 +53,7 @@ public class UserController {
         return new ResponseEntity<>(response, HttpStatus.CREATED);
 
     }
+
 
     @GetMapping("{id}")
     public ResponseEntity<UserModel> findUserById(@PathVariable Long id){
@@ -72,6 +75,11 @@ public class UserController {
     public ResponseEntity<Iterable<UserModel>> findAllUsers(){
 
             Iterable<UserEntity> allUsers = userService.findAllUsers();
+
+            if(!allUsers.iterator().hasNext()){
+                throw new EntityNotFoundException("There are no registered users in the database.");
+            }
+
             Iterable<UserModel> allUsersModels = userMapper.allEntityToAllModels(allUsers);
 
             return new ResponseEntity<>(allUsersModels, HttpStatus.FOUND);
@@ -79,19 +87,29 @@ public class UserController {
     }
 
     @PutMapping()
-    public ResponseEntity<UserModel>updateUser(@RequestBody UserModel user){
+    public ResponseEntity<UserModel>updateUser(@RequestBody UserModel user) {
+
+        if(user.getId()==null){
+            throw new MethodNotAllowedException("You need to specify ID on user to be updated");
+        }
+        if(userService.findUserById(user.getId()).isEmpty()){
+            throw new EntityNotFoundException(responseMessage(user.getId()));
+        }
+
         UserEntity updatedUser = userService.updateUser(userMapper.userModelToUserEntity(user));
         UserModel response = userMapper.userEntityToUserModel(updatedUser);
 
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
-
-
-
     @PatchMapping("{id}")
     public ResponseEntity<UserModel> updateUserEmail(@PathVariable Long id, @RequestBody UserModel user){
+
         Optional<UserEntity> updatedUser = userService.updateUserEmail(id, user.getEmail());
+
+        if(updatedUser.isEmpty()){
+            throw new EntityNotFoundException(responseMessage(id));
+        }
         UserModel response = userMapper.userEntityToUserModel(updatedUser.get());
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
@@ -99,11 +117,19 @@ public class UserController {
 
     @DeleteMapping("{id}")
     public ResponseEntity<Void> deleteUser(@PathVariable Long id){
+
+        if(userService.findUserById(id).isEmpty()){
+            throw new EntityNotFoundException(responseMessage(id));
+        }
+
         userService.deleteUser(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
 
+    private String responseMessage(Long id) {
+        return "There is no user with ID " + id + " in database.";
+    }
 
 
 }
